Removed connection_state_api
