+++ include/boost/mysql/detail/engine_stream_adaptor.hpp
@@ -5,8 +5,8 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_MYSQL_DETAIL_STREAM_ADAPTOR_HPP
-#define BOOST_MYSQL_DETAIL_STREAM_ADAPTOR_HPP
+#ifndef BOOST_MYSQL_DETAIL_ENGINE_STREAM_ADAPTOR_HPP
+#define BOOST_MYSQL_DETAIL_ENGINE_STREAM_ADAPTOR_HPP
 
 #include <boost/mysql/error_code.hpp>
